pyproject.toml
[project]
name = "svtplay-cli"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "anthropic",
    "pydantic",
    "click",
    "mutagen",
    "svtplay-dl",
]

[project.scripts]
svtplay-download-missing = "svtplay.cli.download_missing:main"
svtplay-scan = "svtplay.cli.scan:main"
svtplay-diff = "svtplay.cli.diff:main"
svtplay-meta = "svtplay.cli.meta:main"
svtplay-backfill-info = "svtplay.cli.backfill:info"
svtplay-backfill-apply = "svtplay.cli.backfill:apply"
svtplay-search = "svtplay.cli.search:main"
svtplay-episodes = "svtplay.cli.episodes:main"
svtplay-tmdb = "svtplay.cli.tmdb:main"
svtplay-download = "svtplay.cli.download:main"

[tool.hatch.build.targets.wheel]
packages = ["svtplay"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
